styling and component clean up
