+++ src/components/Image/index.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import { Image } from "theme-ui";
+
+export const ProtocolImage = styled(Image)`
+  height: 48px;
+  width: 48px;
+  margin-right: 8px;
+  clip-path: circle(24px at center);
+`;
